requirements.txt
mesop>=1.2.0,<2.0
google-genai>=1.62.0,<2.0
openai>=2.17.0,<3.0
python-pptx>=1.0.2,<2.0
python-docx>=1.2.0,<2.0
EbookLib>=0.20,<1.0
beautifulsoup4>=4.14.0,<5.0
python-dotenv>=1.2.0,<2.0
pydantic-settings>=2.0,<3.0
pypdf>=6.6.0,<7.0
reportlab>=4.4.0,<5.0
